+++ package.json
@@ -1,28 +1,35 @@
 {
   "name": "d3-org-chart",
-  "version": "1.1.3",
-  "description": "Organization chart, created with d3",
-  "main": "index.js",
-  "scripts": {
-    "test": "echo \"Error: no test specified\" && exit 1"
-  },
-  "repository": {
-    "type": "git",
-    "url": "git+https://github.com/bumbeishvili/d3-organization-chart.git"
-  },
+  "version": "2.0.1",
+  "description": "Highly customizable org chart, created with d3",
   "keywords": [
     "d3",
+    "d3-module",
+    "d3-org-chart",
     "org-chart",
     "hierarchy",
     "employee"
   ],
-  "dependencies": {
-    "d3": "^5.16.0"
-  },
-  "author": "David B.",
-  "license": "ISC",
+  "license": "BSD-3-Clause",
+  "main": "build/d3-org-chart.js",
+  "jsnext:main": "index",
   "bugs": {
     "url": "https://github.com/bumbeishvili/d3-organization-chart/issues"
   },
-  "homepage": "https://github.com/bumbeishvili/d3-organization-chart#readme"
+  "homepage": "https://github.com/bumbeishvili/d3-organization-chart#readme",
+  "repository": {
+    "type": "git",
+    "url": "https://github.com/bumbeishvili/d3-organization-chart.git"
+  },
+  "scripts": {
+    "pretest": "rm -rf build && mkdir build && rollup -f umd -n d3 -o build/d3-org-chart.js -- index.js",
+    "test": "tape 'test/**/*-test.js'",
+    "prepublish": "npm run test && uglifyjs build/d3-org-chart.js -c -m -o build/d3-org-chart.min.js",
+    "postpublish": "zip -j build/d3-org-chart.zip -- LICENSE README.md build/d3-org-chart.js build/d3-org-chart.min.js"
+  },
+  "devDependencies": {
+    "rollup": "0.27",
+    "tape": "4",
+    "uglify-js": "2"
+  }
 }
