+++ test/org-chart-test.js
@@ -1,0 +1,7 @@
+var tape = require("tape"),
+    d3 = require("..");
+
+// tape("org() returns the answer to the ultimate question of life, the universe, and everything.", function(test) {
+//   test.equal(d3.orgChart(), 42);
+//   test.end();
+// });
